ai-services/rembg/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
rembg[gpu]==2.0.62
Pillow==11.1.0
numpy==2.2.2
